+++ packages/working_router_lint/lib/main.dart
@@ -2,7 +2,7 @@ import 'package:analysis_server_plugin/plugin.dart';
 import 'package:analysis_server_plugin/registry.dart';
 
 import 'src/assists/remove_location_tree_element.dart';
-import 'src/assists/wrap_with_group.dart';
+import 'src/assists/wrap_with_scope.dart';
 import 'src/assists/wrap_with_shell.dart';
 
 final plugin = _WorkingRouterPlugin();
@@ -14,7 +14,7 @@ class _WorkingRouterPlugin extends Plugin {
   @override
   void register(PluginRegistry registry) {
     registry.registerAssist(RemoveLocationTreeElement.new);
-    registry.registerAssist(WrapWithGroup.new);
+    registry.registerAssist(WrapWithScope.new);
     registry.registerAssist(WrapWithShell.new);
   }
 }